+++ Sources/PeripheryKit/Results/OutputFormatter.swift
@@ -188,6 +188,8 @@ public extension OutputFormat {
             GitHubActionsFormatter.self
         case .githubMarkdown:
             GitHubMarkdownFormatter.self
+        case .gitlabCodeQuality:
+            GitLabCodeQualityFormatter.self
         }
     }
 }