+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 ##### Enhancements
 
 - Added the `--retain-unused-imported-modules` option.
+- Added the `--format gitlab-codemagic` formatting option for GitLabs Code Quality artifact reports
 
 ##### Bug Fixes
 
+- Fix concurrent mutation crash.
 - Fix unused import false-positive for when a declaration's associated types are declared in different module than module that declares the declaration.
 - Fix unused import false-positive for modules that export unindexed modules.
+- Exclude conditionally imported modules from unused import detection, as they may provide symbols for code that was not compiled.
 
 ## 3.2.0 (2025-06-27)